feat: InputView 역할 축소, 수익률 계산을 long 으로 구현
- 수익률이 너무 높으면 overflow 가 나는 문제가 발견되어
int 에서 long 으로 바꿈
- inputView 가 도메인에 있는 클래스 객체를 return 했었는데
이러한 형태는 inputView 에서 너무 많은 역할을 하는 것으로 보여짐.
따라서 inputView 의 return 타입을 바꾸고
inputView 에 있던 재입력받기 기능을 controller 에 위임함.
- Lotto 생성자가 List<LottoNumber> 를 받는데,
LottoFactory 수동생성시 List<LottoNumber>를 받는 것이
좋아보이지 않아서,
LottoFactory 의 createLottoManual 에서 List<Integer> 를 받도록 수정함.

+++ src/main/java/lotto/domain/ResultStatistic.java
@@ -52,11 +52,11 @@ public class ResultStatistic {
 		return this.results;
 	}
 
-	public int calculateRevenueRate(MoneyForLotto moneyForLotto) {
+	public long calculateRevenueRate(MoneyForLotto moneyForLotto) {
 		return (100 * getTotalRevenue() / moneyForLotto.getMoneyForLotto());
 	}
 
-	private int getTotalRevenue() {
+	private long getTotalRevenue() {
 		int totalRevenue = 0;
 
 		for (Rank rank : Rank.values()) {
